Refactor: Modify file structure and add errors
