+++ src/app.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import cors from 'cors';
 import * as dotenv from 'dotenv';
 import cookieParser from 'cookie-parser';
-import ErrorHandler from './utils/errorHandler.js';
+import ErrorHandler from './middleware/errorHandler.js';
 import { appRouter } from './routes/index.route.js';
 
 dotenv.config();
@@ -22,8 +22,6 @@ app.use(express.json());
 app.use(cookieParser());
 app.use('/api', appRouter);
 
-app.use((err, req, res, next) => {
-  ErrorHandler.handle(err, req, res, next);
-});
+app.use(ErrorHandler);
 
 app.listen(process.env.PORT || 3000, () => console.log('Server On💡'));
