+++ src/routes/shop.route.ts
@@ -1,7 +1,7 @@
 import express from 'express';
 import { shopController } from '../containers/shop.container.js';
-import { verifyAccessToken, verifySeller } from '../utils/verifyAuth.js';
-import asyncHandler from '../utils/asyncHandler.js';
+import { verifyAccessToken, verifySeller } from '../middleware/verifyAuth.js';
+import asyncHandler from '../middleware/asyncHandler.js';
 
 export const shopRouter = express.Router();
