+++ src/utils/structs.ts
